+++ src/ServiceControl.Monitoring.SmokeTests.AzureServiceBus/ServiceControl.Monitoring.SmokeTests.AzureServiceBus.csproj
@@ -7,8 +7,8 @@
     <ProjectGuid>{A82A5634-FC28-47FF-BDAA-69ABBE74A6BC}</ProjectGuid>
     <OutputType>Library</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
-    <RootNamespace>ServiceControl.Monitoring.SmokeTests.AzureServiceBusStandard</RootNamespace>
-    <AssemblyName>ServiceControl.Monitoring.SmokeTests.AzureServiceBusStandard</AssemblyName>
+    <RootNamespace>ServiceControl.Monitoring.SmokeTests.AzureServiceBus</RootNamespace>
+    <AssemblyName>ServiceControl.Monitoring.SmokeTests.AzureServiceBus</AssemblyName>
     <TargetFrameworkVersion>v4.6.1</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <Deterministic>true</Deterministic>
@@ -68,9 +68,9 @@
       <Project>{768be338-20f1-4106-a3eb-dce7686fcbf0}</Project>
       <Name>ServiceControl.Monitoring</Name>
     </ProjectReference>
-    <ProjectReference Include="..\ServiceControl.Transports.AzureServiceBusStandard\ServiceControl.Transports.AzureServiceBusStandard.csproj">
-      <Project>{3cda118c-521a-4fd1-bdab-4059d141d6f2}</Project>
-      <Name>ServiceControl.Transports.AzureServiceBusStandard</Name>
+    <ProjectReference Include="..\ServiceControl.Transports.AzureServiceBus\ServiceControl.Transports.AzureServiceBus.csproj">
+      <Project>{37baaa9b-a17b-4cb7-8359-d09586cdaad2}</Project>
+      <Name>ServiceControl.Transports.AzureServiceBus</Name>
     </ProjectReference>
   </ItemGroup>
   <ItemGroup>
