+++ src/ServiceControl.Transports.AzureServiceBus/AzureServiceBusTransport.cs
@@ -1,4 +1,4 @@
-namespace ServiceControl.Transports.AzureServiceBusStandard
+namespace ServiceControl.Transports.AzureServiceBus
 {
     public class AzureServiceBusTransport : NServiceBus.AzureServiceBusTransport
     {
